requirements.txt
# Core dependencies
anthropic==0.39.0
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Development and utilities
colorama==0.4.6
